+++ montepy/data_inputs/material.py
@@ -69,7 +69,7 @@ class _DefaultLibraries:
         try:
             node = self._libraries[key]
         except KeyError:
-            node = self._generate_default_node(key)
+            node = self._generate_default_node(key, str(value))
             self._parent()._append_param_lib(node)
             self._libraries[key] = node
         node["data"].value = str(value)
@@ -80,7 +80,7 @@ class _DefaultLibraries:
         self._parent()._delete_param_lib(node)
 
     def __str__(self):
-        return str(self._libraries)
+        return "\n".join([f"{key} = {value}" for key, value in self.items()])
 
     def __iter__(self):
         return iter(self._libraries)
@@ -98,13 +98,13 @@ class _DefaultLibraries:
         return key
 
     @staticmethod
-    def _generate_default_node(key: LibraryType):
+    def _generate_default_node(key: LibraryType, val: str):
         classifier = syntax_node.ClassifierNode()
         classifier.prefix = key.value
         ret = {
             "classifier": classifier,
             "seperator": syntax_node.PaddingNode(" = "),
-            "data": syntax_node.ValueNode("", str),
+            "data": syntax_node.ValueNode(val, str),
         }
         return syntax_node.SyntaxNode("mat library", ret)
 
@@ -286,6 +286,7 @@ class Material(data_input.DataInputAbstract, Numbered_MCNP_Object):
     """
 
     _parser = MaterialParser()
+    _NEW_LINE_STR = "\n" + " " * DEFAULT_INDENT
 
     def __init__(
         self,
@@ -372,6 +373,7 @@ class Material(data_input.DataInputAbstract, Numbered_MCNP_Object):
 
         This is called from _DefaultLibraries.
         """
+        self._ensure_has_ending_padding()
         self._tree["data"].append_param(node)
 
     def _delete_param_lib(self, node: syntax_node.SyntaxNode):
@@ -641,15 +643,42 @@ See <https://www.montepy.org/migrations/migrate0_1.html> for more information ""
         self._check_valid_comp(nuclide_frac_pair)
         self._elements.add(nuclide_frac_pair[0].element)
         self._nuclei.add(nuclide_frac_pair[0].nucleus)
+        # node for fraction
         node = self._generate_default_node(
-            float, str(nuclide_frac_pair[1]), "\n" + " " * DEFAULT_INDENT
+            float, str(nuclide_frac_pair[1]), self._NEW_LINE_STR
         )
         syntax_node.ValueNode(str(nuclide_frac_pair[1]), float)
         node.is_negatable_float = True
         nuclide_frac_pair = (nuclide_frac_pair[0], node)
         node.is_negative = not self._is_atom_fraction
         self._components.append(nuclide_frac_pair)
+        self._ensure_has_ending_padding()
         self._tree["data"].append_nuclide(("_", nuclide_frac_pair[0]._tree, node))
+
+    def _ensure_has_ending_padding(self):
+        def get_last_val_node():
+            last_vals = self._tree["data"].nodes[-1][-1]
+            if isinstance(last_vals, syntax_node.ValueNode):
+                return last_vals
+            return last_vals["data"]
+
+        if len(self._tree["data"]) == 0:
+            return
+        padding = get_last_val_node().padding
+
+        def add_new_line_padding():
+            if padding is None:
+                get_last_val_node().padding = syntax_node.PaddingNode(
+                    self._NEW_LINE_STR
+                )
+            else:
+                padding.append(self._NEW_LINE_STR)
+
+        if padding:
+            padding.check_for_graveyard_comments(True)
+            add_new_line_padding()
+        else:
+            add_new_line_padding()
 
     def change_libraries(self, new_library: Union[str, Library]):
         """
@@ -805,6 +834,13 @@ See <https://www.montepy.org/migrations/migrate0_1.html> for more information ""
                 all(element_search.values()),
             )
         )
+
+    def clear(self):
+        """
+        Clears all nuclide components from this material.
+        """
+        for _ in range(len(self)):
+            del self[0]
 
     def normalize(self):
         """